+++ assistants/ai/openai.py
@@ -105,14 +105,14 @@ class Assistant(
     REASONING_MODELS = REASONING_MODELS
 
     def __init__(  # pylint: disable=too-many-arguments
-            self,
-            *,
-            name: str,
-            model: str,
-            instructions: str,
-            tools: list | NotGiven = NOT_GIVEN,
-            api_key: str = environment.OPENAI_API_KEY,
-            thinking: ThinkingLevel = 1,
+        self,
+        *,
+        name: str,
+        model: str,
+        instructions: str,
+        tools: list | NotGiven = NOT_GIVEN,
+        api_key: str = environment.OPENAI_API_KEY,
+        thinking: ThinkingLevel = 1,
     ):
         """
         Initialize the Assistant instance.
@@ -191,7 +191,7 @@ class Assistant(
 
         # Add system message if instructions are available
         if self.instructions and not any(
-                msg.get("role") == "system" for msg in self.memory
+            msg.get("role") == "system" for msg in self.memory
         ):
             input_messages.append({"role": "system", "content": self.instructions})
 
@@ -234,7 +234,7 @@ class Assistant(
         return response.data[0].url
 
     async def converse(
-            self, user_input: str, thread_id: Optional[str] = None
+        self, user_input: str, thread_id: Optional[str] = None
     ) -> Optional[MessageData]:
         """
         Converse with the assistant by sending a message and getting a response.
@@ -277,11 +277,11 @@ class Completion(ReasoningModelMixin, ConversationHistoryMixin, AssistantInterfa
     REASONING_MODELS = REASONING_MODELS
 
     def __init__(
-            self,
-            model: str,
-            max_tokens: int = 4096,
-            api_key: str = environment.OPENAI_API_KEY,
-            thinking: ThinkingLevel = 2,
+        self,
+        model: str,
+        max_tokens: int = 4096,
+        api_key: str = environment.OPENAI_API_KEY,
+        thinking: ThinkingLevel = 2,
     ):
         """
         Initialize the Completion instance.
@@ -330,7 +330,7 @@ class Completion(ReasoningModelMixin, ConversationHistoryMixin, AssistantInterfa
         return response.choices[0].message
 
     async def converse(
-            self, user_input: str, thread_id: Optional[str] = None
+        self, user_input: str, thread_id: Optional[str] = None
     ) -> Optional[MessageData]:
         """
         Converse with the assistant using the chat completion API.
@@ -366,13 +366,17 @@ class Completion(ReasoningModelMixin, ConversationHistoryMixin, AssistantInterfa
         temp_memory = deepcopy(self.memory)
         if (message := temp_memory[0])["role"] == "system":
             if "You should always respond in audio format." not in message["content"]:
-                message["content"] = f"""\
+                message[
+                    "content"
+                ] = f"""\
 You should always respond in audio format.
 
 {message["content"]}
 """
         complete = False
         while not complete:
+            if all(message.get("audio") is None for message in temp_memory):
+                temp_memory = [message]
             try:
                 completion = self.client.chat.completions.create(
                     model="gpt-4o-audio-preview",
@@ -395,10 +399,27 @@ You should always respond in audio format.
         response = completion.choices[0].message
 
         if response.audio:
-            self.remember({"role": "assistant", "audio": {"id": response.audio.id},
-                           "content": f"[AUDIO TRANSCRIPTION]: {response.content}"})
+            self.remember(
+                {
+                    "role": "assistant",
+                    "audio": {"id": response.audio.id},
+                    "content": f"[AUDIO TRANSCRIPTION]: {response.content}",
+                }
+            )
             return base64.b64decode(completion.choices[0].message.audio.data)
+        else:
+            self.remember({"role": "assistant", "content": response.content})
+
         return response.content
 
 
-# e = {'error': {'message': "Invalid voice: 'nova'. A previous message you provided used a different voice 'ballad'. Only one voice may be used throughout a conversation.", 'type': 'invalid_request_error', 'param': 'audio.voice', 'code': 'invalid_voice'}}
+class RealtimeVoiceChat(ConversationHistoryMixin, AssistantInterface):
+    async def start(self) -> None:
+        pass
+
+    async def converse(
+        self, user_input: str, thread_id: Optional[str] = None
+    ) -> Optional[MessageData]:
+        filename = user_input
+        with open(filename, "rb") as f:
+            audio = f.read()