+++ assistants/config/__init__.py
@@ -4,7 +4,10 @@ from argparse import Namespace
 from dataclasses import dataclass, asdict
 from enum import Enum
 
-from assistants.lib.constants import CLAUDE_CLI_MAX_TOKENS
+from assistants.lib.constants import (
+    DEFAULT_MAX_HISTORY_TOKENS,
+    DEFAULT_MAX_RESPONSE_TOKENS,
+)
 
 
 class CustomKeyNames(str, Enum):
@@ -79,10 +82,10 @@ def get_config() -> Config:
         ASSISTANT_NAME=os.environ.get("ASSISTANT_NAME", "DefaultAssistant"),
         TELEGRAM_BOT_TOKEN=os.environ.get("TG_BOT_TOKEN", None),
         DEFAULT_MAX_HISTORY_TOKENS=os.environ.get(
-            "DEFAULT_MAX_TOKENS", CLAUDE_CLI_MAX_TOKENS
+            "DEFAULT_MAX_TOKENS", DEFAULT_MAX_HISTORY_TOKENS
         ),
         DEFAULT_MAX_RESPONSE_TOKENS=os.environ.get(
-            "DEFAULT_MAX_RESPONSE_TOKENS", CLAUDE_CLI_MAX_TOKENS
+            "DEFAULT_MAX_RESPONSE_TOKENS", DEFAULT_MAX_RESPONSE_TOKENS
         ),
         OPEN_IMAGES_IN_BROWSER=bool(
             json.loads(os.environ.get("OPEN_IMAGES_IN_BROWSER", "true"))