+++ assistants/lib/constants.py
@@ -20,4 +20,5 @@ Press Ctrl+C or Ctrl+D to exit the program
 Anything else you type will be sent to the assistant for processing.\
 """
 
-CLAUDE_CLI_MAX_TOKENS = 4096
+DEFAULT_MAX_HISTORY_TOKENS = 10000
+DEFAULT_MAX_RESPONSE_TOKENS = 4096